+++ src/saveload/plans_sl.cpp
@@ -28,10 +28,10 @@ static const SaveLoad _plan_desc[] = {
 static void RealSave_PLAN(Plan *p)
 {
 	SlObject(p, _plan_desc);
-	SlWriteUint32(p->lines.size());
+	SlWriteUint32((uint32)p->lines.size());
 	for (size_t i = 0; i < p->lines.size(); i++) {
 		PlanLine *pl = p->lines[i];
-		SlWriteUint32(pl->tiles.size());
+		SlWriteUint32((uint32)pl->tiles.size());
 		SlArray(&pl->tiles[0], pl->tiles.size(), SLE_UINT32);
 	}
 }
